requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
Pillow>=10.0
openai>=1.0
anthropic>=0.30
pyyaml>=6.0
python-dotenv>=1.0
